added disable-server flag

+++ B/features/fixture.py
@@ -1,6 +1,7 @@
 import subprocess
 import requests
 from behave import fixture
+import sys
 
 def set_up_and_tear_down():
     # Set up make sure the app is not running at first
@@ -24,14 +25,18 @@ def set_up_and_tear_down():
 
 @fixture(name="fixture.app")
 def app(context):
-    try:
-        process = subprocess.Popen(["java", "-jar", "runTodoManagerRestAPI-1.5.5.jar"], shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT)
-        status_code = subprocess.call(['curl', 'http://localhost:4567'], shell=True)
+    #  we simply don't start the jar if we dont wan't the runTodoManagerRestAPI to run.
+    if "--disable-server" in sys.argv:
+        yield
+    else:
+        try:
+            process = subprocess.Popen(["java", "-jar", "runTodoManagerRestAPI-1.5.5.jar"], shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT)
+            status_code = subprocess.call(['curl', 'http://localhost:4567'], shell=True)
 
-        while status_code:  # Verify the system is up before starting the test
-            status_code = subprocess.call(['curl', 'http://localhost:4567'], shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT)
-        yield process
-    finally:
-        subprocess.call(['curl', 'http://localhost:4567/shutdown'], shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT)
-        process.kill()
+            while status_code:  # Verify the system is up before starting the test
+                status_code = subprocess.call(['curl', 'http://localhost:4567'], shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT)
+            yield process
+        finally:
+            subprocess.call(['curl', 'http://localhost:4567/shutdown'], shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT)
+            process.kill()
